requirements.txt
snowflake-connector-python==2.7.2
dbt-snowflake==1.0.0
airflow-dbt==0.4.0
apache-airflow-providers-http
apache-airflow-providers-airbyte
apache-airflow-providers-slack
apache-airflow-providers-slack[http]


# snowflake-connector-python[pandas]
# shutil